requirements.txt
ExifRead==3.0.0
Pillow==9.5.0
requests==2.28.2
torch==2.0.0
torchvision==0.15.1
hydra-core==1.3.2
omegaconf==2.3.0
regex==2023.3.23
pip install torch
!pip install numpy
!pip install cython
!pip install -r https://raw.githubusercontent.com/pytorch/fairseq/master/requirements.txt